// src/api/types/TranscribeTranscriptData.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export interface TranscribeTranscriptData {
    /** Transcript segment with punctuations applied and command phrases removed */
    text: string;
    /** The raw transcript without spoken punctuation applied and without command phrases removed */
    rawTranscriptText: string;
    /** Start time of the transcript segment in seconds */
    start: number;
    /** End time of the transcript segment in seconds */
    end: number;
    /** If false, then interim transcript result */
    isFinal: boolean;
}